# Schefco.TaskFlow.API/bin/Debug/net8.0/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TaskFlow API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7154"
    },
    {
      "url": "https://taskflow-backend-96ki.onrender.com"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User registered successfully" },
          "400": { "description": "Invalid registration request" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user and set HttpOnly cookie",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. AuthToken cookie is set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get the authenticated user's identity from the HttpOnly cookie",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Authenticated user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserIdentity"
                }
              }
            }
          },
          "401": { "description": "User is not authenticated" }
        }
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "summary": "Reset a user's password and return a new temporary password",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserPasswordCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temporary password generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetUserPasswordResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": { "type": "string" }
        },
        "required": [ "email", "password" ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": { "type": "string" }
        },
        "required": [ "email", "password" ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "requiresPasswordReset": { "type": "boolean" }
        },
        "required": [ "requiresPasswordReset" ]
      },
      "UserIdentity": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": { "type": "string" },
          "role": { "type": "string" }
        },
        "required": [ "userId", "email", "role" ]
      },
      "ResetUserPasswordCommand": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [ "userId" ]
      },
      "ResetUserPasswordResponse": {
        "type": "object",
        "properties": {
          "tempPassword": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {}
  }
}